make artifact tree fetch its own data
