+++ scopes/component/ui/component-artifacts/artifacts-tree/artifacts-tree.module.scss
@@ -47,6 +47,9 @@
 
 .link {
   color: var(--bit-accent-color, #6c5ce7);
+}
+
+.node {
   pointer-events: none;
   > div {
     pointer-events: auto;
